fix: move cache tests to upload_tests.rs (file-size gate) + clippy sort_by_key
